@@ -1,40 +1,36 @@
 package fx
 
-type fmap[T, U any] struct {
-	it Iterator[T]
-	fn func(v T) (U, bool)
-	v  U
-}
+import "iter"
 
-func (f *fmap[T, U]) Value() U {
-	return f.v
-}
-
-func (f *fmap[T, U]) Next() bool {
-	for {
-		if !f.it.Next() {
-			var v U
-			f.v = v
-			return false
-		}
-		if v, ok := f.fn(f.it.Value()); ok {
-			f.v = v
-			return true
-		}
-	}
-}
-
-func OfType[T, U any](it Iterator[T]) Iterator[U] {
-	return FMap[T, U](it, func(v T) (U, bool) {
-		u, ok := ((interface{})(v)).(U)
+// OfType returns a sequence composed of all elements in the input sequence that are of type U.
+func OfType[T, U any](it iter.Seq[T]) iter.Seq[U] {
+	return FMap(it, func(v T) (U, bool) {
+		u, ok := any(v).(U)
 		return u, ok
 	})
 }
 
-func FMap[T, U any](it Iterator[T], fn func(v T) (U, bool)) Iterator[U] {
-	return &fmap[T, U]{it: it, fn: fn}
+// FMap returns a sequence of values computed by invoking fn on each element of the input sequence and returning only
+// mapped values for with fn returns true.
+func FMap[T, U any](it iter.Seq[T], fn func(v T) (U, bool)) iter.Seq[U] {
+	return func(yield func(v U) bool) {
+		for v := range it {
+			if u, ok := fn(v); ok {
+				if !yield(u) {
+					return
+				}
+			}
+		}
+	}
 }
 
-func Map[T, U any](it Iterator[T], fn func(v T) U) Iterator[U] {
-	return FMap[T, U](it, func(v T) (U, bool) { return fn(v), true })
+// Map invokes fn on each value in the input sequence and returns the results.
+func Map[T, U any](it iter.Seq[T], fn func(v T) U) iter.Seq[U] {
+	return func(yield func(v U) bool) {
+		for v := range it {
+			if !yield(fn(v)) {
+				return
+			}
+		}
+	}
 }
