@@ -1,28 +1,15 @@
 package fx
 
-type take[T any] struct {
-	n  int
-	it Iterator[T]
-	v  T
-}
+import "iter"
 
-func (t *take[T]) Value() T {
-	return t.v
-}
-
-func (t *take[T]) Next() bool {
-	if t.n <= 0 {
-		return false
+// Take returns an iterator that takes at most n values from its source.
+func Take[T any](it iter.Seq[T], n int) iter.Seq[T] {
+	return func(yield func(v T) bool) {
+		for v := range it {
+			if n <= 0 || !yield(v) {
+				return
+			}
+			n--
+		}
 	}
-	if !t.it.Next() {
-		var v T
-		t.v = v
-		return false
-	}
-	t.v, t.n = t.it.Value(), t.n-1
-	return true
-}
-
-func Take[T any](it Iterator[T], n int) Iterator[T] {
-	return &take[T]{n: n, it: it}
 }
