@@ -2,6 +2,8 @@ package fx
 
 import (
 	"fmt"
+	"iter"
+	"slices"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -9,7 +11,7 @@ import (
 
 func TestTake(t *testing.T) {
 	cases := []struct {
-		it       Iterator[int]
+		it       iter.Seq[int]
 		n        int
 		expected []int
 	}{
@@ -26,7 +28,7 @@ func TestTake(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(fmt.Sprintf("%v", c.expected), func(t *testing.T) {
-			actual := ToSlice(Take(c.it, c.n))
+			actual := slices.Collect(Take(c.it, c.n))
 			assert.Equal(t, c.expected, actual)
 		})
 	}
