@@ -1,4 +1,4 @@
-package fx
+package try
 
 import (
 	"iter"
@@ -6,16 +6,16 @@ import (
 )
 
 // Concat returns an iterator that returns values from each iterator in sequence.
-func Concat[T any](iters ...iter.Seq[T]) iter.Seq[T] {
+func Concat[T any](iters ...iter.Seq2[T, error]) iter.Seq2[T, error] {
 	return ConcatMany(slices.Values(iters))
 }
 
 // ConcatMany returns an iterator that returns values from each iterator in sequence.
-func ConcatMany[T any](iters iter.Seq[iter.Seq[T]]) iter.Seq[T] {
-	return func(yield func(v T) bool) {
+func ConcatMany[T any](iters iter.Seq[iter.Seq2[T, error]]) iter.Seq2[T, error] {
+	return func(yield func(v T, err error) bool) {
 		for it := range iters {
-			for v := range it {
-				if !yield(v) {
+			for v, err := range it {
+				if !yield(v, err) {
 					return
 				}
 			}
